EI-423-fix-period-number

+++ src/SFA.DAS.EmployerIncentives.Database/Tables/CollectionCalendar.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [incentives].[CollectionCalendar]
 (
 	[Id] INT NOT NULL PRIMARY KEY NONCLUSTERED,
-	[PeriodNumber] SMALLINT NOT NULL,	
+	[PeriodNumber] TINYINT NOT NULL,	
 	[CalendarMonth] TINYINT NOT NULL,
 	[CalendarYear] SMALLINT NOT NULL,
 	[EIScheduledOpenDateUTC] DATETIME2 NOT NULL,
